test: move the provenance tests into the feature-central file
`test_multimodal_tool_returns.py` is where this feature's cross-provider
matrix lives, and where `MARKER_RECORDED` says which cells are backed by a
recording and which lean on a mapper test instead. The two mapper tests that
stand in for the unrecorded Groq and xAI cells were sitting 6,000 lines deep in
their provider monoliths, so the note and the tests it depends on could not be
read together.

Move both, plus the cross-provider replay test, which was the only thing in
`test_model.py` constructing provider models and snapshotting wire shapes --
its two neighbours there really are `prepare_messages` tests and stay put.
